% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_plot.R
\name{plot.dot_plot}
\alias{plot.dot_plot}
\title{plot methods for dot_plot object}
\usage{
\method{plot}{dot_plot}(x, ...)
}
\arguments{
\item{x}{dot_plot object}

\item{...}{other arguments for generic methods}
}
\description{
plot methods for dot_plot object
}
