% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_plot.R
\name{print.dot_plot}
\alias{print.dot_plot}
\title{print methods for dot_plot object}
\usage{
\method{print}{dot_plot}(x, ...)
}
\arguments{
\item{x}{dot_plot object}

\item{...}{other arguments for generic methods}
}
\description{
print methods for dot_plot object
}
