% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_summary.R
\name{print.safety_summary}
\alias{print.safety_summary}
\title{print method for safety summary object}
\usage{
\method{print}{safety_summary}(x, ...)
}
\arguments{
\item{x}{a safety_summary object}

\item{...}{extra arguments for the generic print method}
}
\description{
print method for safety summary object
}
\keyword{internal}
